service_provider_project/requirements.txt
Django>=4.0,<5.0
requests>=2.31.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
cryptography>=42.0.0
psycopg2-binary>=2.9.9